# 🧪 Improve test coverage and testing infrastructure (Issue #6)