--- a/mix.exs
+++ b/mix.exs
@@ -8,7 +8,15 @@ defmodule Grapple.MixProject do
       elixir: "~> 1.18",
       start_permanent: Mix.env() == :prod,
       deps: deps(),
-      docs: docs()
+      docs: docs(),
+      test_coverage: [tool: ExCoveralls],
+      preferred_cli_env: [
+        coveralls: :test,
+        "coveralls.detail": :test,
+        "coveralls.post": :test,
+        "coveralls.html": :test,
+        "coveralls.json": :test
+      ]
     ]
   end
 
@@ -26,7 +34,10 @@ defmodule Grapple.MixProject do
       {:jason, "~> 1.4"},
       {:ex_doc, "~> 0.31", only: :dev, runtime: false},
       {:benchee, "~> 1.3", only: :dev, runtime: false},
-      {:benchee_html, "~> 1.0", only: :dev, runtime: false}
+      {:benchee_html, "~> 1.0", only: :dev, runtime: false},
+      {:stream_data, "~> 1.1", only: :test},
+      {:excoveralls, "~> 0.18", only: :test},
+      {:mox, "~> 1.1", only: :test}
     ]
   end