--- a/test/test_helper.exs
+++ b/test/test_helper.exs
@@ -1,1 +1,9 @@
-ExUnit.start()
+# Configure ExUnit
+ExUnit.start(
+  exclude: [:integration],
+  formatters: [ExUnit.CLIFormatter],
+  max_cases: System.schedulers_online() * 2
+)
+
+# Start the application for integration tests
+Application.ensure_all_started(:grapple)